+++ adapter/bidmachine/src/main/java/org/bidon/bidmachine/impl/BMBannerAdImpl.kt
@@ -52,6 +52,7 @@ internal class BMBannerAdImpl(
                         BidType.CPM -> {
                             setNetworks("")
                             setPlacementId(adParams.placement)
+                            setTargetingParams(adParams.targetingParams)
                         }
 
                         BidType.RTB -> {