+++ adapter/bidmachine/src/main/java/org/bidon/bidmachine/impl/BMRewardedAdImpl.kt
@@ -56,6 +56,7 @@ internal class BMRewardedAdImpl(
                     BidType.CPM -> {
                         setNetworks("")
                         setPlacementId(adParams.placement)
+                        setTargetingParams(adParams.targetingParams)
                     }
 
                     BidType.RTB -> {