+++ adapter/bidmachine/src/main/java/org/bidon/bidmachine/impl/GetAdAuctionParamUseCase.kt
@@ -1,6 +1,7 @@
 package org.bidon.bidmachine.impl
 
 import io.bidmachine.CustomParams
+import io.bidmachine.TargetingParams
 import org.bidon.bidmachine.BMBannerAuctionParams
 import org.bidon.bidmachine.BMFullscreenAuctionParams
 import org.bidon.sdk.adapter.AdAuctionParamSource
@@ -12,48 +13,74 @@ import org.json.JSONObject
 internal class GetAdAuctionParamUseCase {
     fun getBMFullscreenAuctionParams(auctionParamsScope: AdAuctionParamSource): Result<BMFullscreenAuctionParams> {
         return auctionParamsScope {
+            val extra = adUnit.extra
             BMFullscreenAuctionParams(
                 price = adUnit.pricefloor,
                 timeout = adUnit.timeout,
                 context = activity.applicationContext,
                 adUnit = adUnit,
-                payload = adUnit.extra?.optString("payload"),
-                placement = adUnit.extra?.optString("placement"),
-                customParameters = buildCustomParameters(adUnit.extra)
+                customParameters = extra.buildCustomParameters(),
+                targetingParams = extra.buildTargetingParams(),
+                payload = extra?.optString("payload"),
+                placement = extra?.optString("placement"),
             )
         }
     }
 
     fun getBMBannerAuctionParams(auctionParamsScope: AdAuctionParamSource): Result<BMBannerAuctionParams> {
         return auctionParamsScope {
+            val extra = adUnit.extra
             BMBannerAuctionParams(
                 price = adUnit.pricefloor,
                 timeout = adUnit.timeout,
                 activity = activity,
                 bannerFormat = bannerFormat,
                 adUnit = adUnit,
-                payload = adUnit.extra?.optString("payload"),
-                placement = adUnit.extra?.optString("placement"),
-                customParameters = buildCustomParameters(adUnit.extra),
+                customParameters = extra.buildCustomParameters(),
+                targetingParams = extra.buildTargetingParams(),
+                payload = extra?.optString("payload"),
+                placement = extra?.optString("placement"),
             )
         }
     }
 
-    private fun buildCustomParameters(extra: JSONObject?): CustomParams {
-        val customParams = CustomParams().addParam("mediation_mode", "bidon")
-        val paramsJson = extra?.optJSONObject("custom_parameters") ?: return customParams
-
-        try {
-            val keys = paramsJson.keys()
-            while (keys.hasNext()) {
-                val key = keys.next()
-                val value = paramsJson.optString(key)
-                customParams.addParam(key, value)
+    private fun JSONObject?.buildTargetingParams(): TargetingParams {
+        return TargetingParams().apply {
+            this@buildTargetingParams?.toList("bcat")?.forEach { category ->
+                addBlockedAdvertiserIABCategory(category)
             }
-        } catch (_: Throwable) {
-            // ignore
+            this@buildTargetingParams?.toList("badv")?.forEach { domain ->
+                addBlockedAdvertiserDomain(domain)
+            }
+            this@buildTargetingParams?.toList("bapps")?.forEach { app ->
+                addBlockedApplication(app)
+            }
         }
+    }
 
-        return customParams
+    private fun JSONObject?.buildCustomParameters(): CustomParams {
+        return CustomParams().apply {
+            addParam("mediation_mode", "bidon")
+
+            this@buildCustomParameters?.optJSONObject("custom_parameters")?.let { paramsJson ->
+                paramsJson.keys().forEach { key ->
+                    paramsJson.optString(key).takeIf { it.isNotEmpty() }?.let { value ->
+                        addParam(key, value)
+                    }
+                }
+            }
+        }
+    }
+
+    private fun JSONObject.toList(key: String): List<String>? {
+        return optJSONArray(key)?.let { jsonArray ->
+            buildList {
+                repeat(jsonArray.length()) { index ->
+                    jsonArray.optString(index).takeIf { it.isNotEmpty() }?.let { value ->
+                        add(value)
+                    }
+                }
+            }.takeIf { it.isNotEmpty() }
+        }
     }
 }