+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Features:
 - [BDN-1081](https://appodeal.atlassian.net/browse/BDN-1081) Fixed incorrect bid status marked as SUCCESSFUL when bid.price < pricefloor
 - [BDN-1071](https://appodeal.atlassian.net/browse/BDN-1071) Implemented RTB mode for Yandex adapter
+- [BDN-1076](https://appodeal.atlassian.net/browse/BDN-1076) Added support for IAB restriction parameters (bcat, badv, bapps) for BidMachine CPM ad units
 
 # 0.12.0 (2025.11.03)
 ## Features: